+++ src/nb/nbayes.rs
@@ -136,7 +136,7 @@ pub fn classify(
         let class_id_opt = &models.iter().position(|m| m.class_name == seq.class_name);
         if let Some(class_id) = *class_id_opt {
             let probs: Vec<f64> = models.iter().map(|m| m.log_prob_sequence(&seq)).collect();
-            c12n.add_case(class_id, probs, show_ranked, || {
+            c12n.add_case(class_id, &seq.class_name, probs, show_ranked, || {
                 format!("\n{}: '{}'\n", filename, seq.class_name)
             });
         }